chore: rename dateFormatters to dateFormatter for better semantix and fix class export for dateFormatter
